refactor(backend): extract shortform generation and compose actions
